steem DatabaseApi update
