+++ Sources/Ditch.Steem/OperationManager.UnsortedApi.cs
@@ -1,0 +1,47 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Threading;
+using Ditch.Core.JsonRpc;
+using Ditch.Steem.Operations;
+using Ditch.Steem.Operations.Enums;
+using Ditch.Steem.Operations.Get;
+using Newtonsoft.Json.Linq;
+
+namespace Ditch.Steem
+{
+    public partial class OperationManager
+    {
+        /// <summary>
+        /// 
+        /// Возвращает список: Либо всех подписчиков пользователя 'following'. 
+        /// Либо если указано имя пользователя в параметре 'startFollower' возвращается список совпадающих подписчиков.
+        /// </summary>
+        /// <param name="following"></param>
+        /// <param name="startFollower"></param>
+        /// <param name="followType"></param>
+        /// <param name="limit"></param>
+        /// <param name="token">Throws a <see cref="T:System.OperationCanceledException" /> if this token has had cancellation requested.</param>
+        /// <returns></returns>
+        /// <exception cref="T:System.OperationCanceledException">The token has had cancellation requested.</exception>
+        public JsonRpcResponse<FollowApiObj[]> GetFollowers(string following, string startFollower, FollowType followType, UInt16 limit, CancellationToken token)
+        {
+            return CustomGetRequest<FollowApiObj[]>("call", token, "follow_api", "get_followers", new object[] { following, startFollower, followType.ToString().ToLower(), limit });
+        }
+
+        /// <summary>
+        /// 
+        /// Aналогично GetFollowers только для подписок
+        /// </summary>
+        /// <param name="follower"></param>
+        /// <param name="startFollowing"></param>
+        /// <param name="followType"></param>
+        /// <param name="limit"></param>
+        /// <param name="token">Throws a <see cref="T:System.OperationCanceledException" /> if this token has had cancellation requested.</param>
+        /// <returns></returns>
+        /// <exception cref="T:System.OperationCanceledException">The token has had cancellation requested.</exception>
+        public JsonRpcResponse<FollowApiObj[]> GetFollowing(string follower, string startFollowing, FollowType followType, UInt16 limit, CancellationToken token)
+        {
+            return CustomGetRequest<FollowApiObj[]>("call", token, "follow_api", "get_following", new object[] { follower, startFollowing, followType.ToString().ToLower(), limit });
+        }
+    }
+}
